auto_generate_video_fold6/services/storage-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
minio==7.2.0
boto3==1.34.0
botocore==1.34.0
Pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1